requirements.txt
annotated-types==0.6.0
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
defusedxml==0.7.1
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
humanfriendly==10.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
onnx==1.16.0
onnxruntime==1.17.3
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
protobuf==5.26.1
pycocotools==2.0.7
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.1.2
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
scipy==1.13.0
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
six==1.16.0
supervision==0.19.0
sympy==1.12
torch==2.2.2+cu118
torchaudio==2.2.2+cu118
torchvision==0.17.2+cu118
typing_extensions==4.11.0
